i think a working component system? idrk what i'm doing atp

+++ modules/camera/camera.c
@@ -6,7 +6,7 @@ void component_camera_start(component_t* self) {
 	console_write("Camera Start Called\n");
 }
 
-void component_camera_update(component_t* self, float delta) {
+void component_camera_update(component_t* self, double delta) {
 	(void)self;
 	(void)delta;
 }
