+++ modules/camera/camera.h
@@ -30,6 +30,6 @@ typedef struct {
 } component_camera_t;
 
 void component_camera_start(component_t* self);
-void component_camera_update(component_t* self, float delta);
+void component_camera_update(component_t* self, double delta);
 void component_camera_post_render(component_t* self);
 void component_camera_shutdown(component_t* self);
